[FIX] assets_management: Apply old migration
Fields `from_nr` and `to_nr` should have been migrated into `from_year_nr` and `to_year_nr` respectively.
In DBs that did not perform the migration at that time, data is lost.
Restore the data of the only line we have in the module's data files.

Also remove the original migration file that was lost in the repo.